+++ assets/javascripts/kitten/components/popovers/simple-popover/index.js
@@ -6,7 +6,7 @@ import { CrossIcon } from '../../../components/icons/cross-icon'
 import COLORS from '../../../constants/colors-config'
 import { pxToRem } from '../../../helpers/utils/typography'
 import { Title } from '../../../components/typography/title'
-import { Paragraph } from '../../../components/typography/paragraph'
+import { Paragraph } from '../../../components/typography/paragraph/next'
 import { CONTAINER_PADDING_THIN } from '../../../constants/grid-config'
 import { ScreenConfig } from '../../../constants/screen-config'
 import { Button } from '../../../components/buttons/button/button'
@@ -135,7 +135,7 @@ export const SimplePopover = ({
           </Title>
         </Marger>
         <Marger top=".5" bottom="1">
-          <Paragraph modifier="quaternary">{text}</Paragraph>
+          <Paragraph modifier="tertiary">{text}</Paragraph>
         </Marger>
         {buttons.length > 0 && (
           <ButtonsContainer top="3">